Updated the ImageSource construct to be more adaptable. Added interlaced and aggregated ImageSource objects to extract this functionality from the PeakFit plugin. The ImageSource is now transparent to the consumer objects. The ImageSource can describe the parent image source, original image source for the data and the start and end frames of the data that were returned when reading the source.
PeakFit results from aggregated data are now labelled with the start and
end frame of all the frames that were aggregated. Thus the exposure time
does not have to be adjusted.

Added JUnit tests for the interlaced and aggregated ImageSource
constructs. Added a MemoryImageSource to allow testing using a simple
image stack.