Added deviations estimate to the MLE fitted parameters. This is derived from the Fisher Information matrix and assumes the fit has achieved the Cramer-Roa lower bounds.
Note that the deviations from the least squares estimators are lower
than the deviations from the Fisher matrix. However the Apache LVM
fitter and the custom NonLinearFit class produce the same output so this
is presumed correct, it is just that the fitters have a smaller
confidence interval on the fitted parameters than the Cramer-Roa lower
bounds.